server/requirements.txt
python_dotenv
flask
flask_cors
flask_jwt_extended
flask_bcrypt
flask_sqlalchemy
pytz
numpy==2.2.3
torch==2.0.1
torchvision==0.15.2
opencv-python
pillow
tqdm